+++ src/finance/programgarden_finance/ls/overseas_stock/real/AS0/blocks.py
@@ -453,19 +453,19 @@ class AS0RealResponseBody(BaseModel):
         description="Parent order number for grouped orders; consume as returned by LS.",
         examples=["0"],
     )
-    sOrgOrdUnercQty: int = Field(
+    sOrgOrdUnercQty: float = Field(
         ...,
         title="원주문미체결수량 (Original-order unfilled quantity)",
         description="Unfilled quantity remaining on the original order.",
         examples=[0, 5],
     )
-    sOrgOrdMdfyQty: int = Field(
+    sOrgOrdMdfyQty: float = Field(
         ...,
         title="원주문정정수량 (Original-order modified quantity)",
         description="Quantity that has been modified on the original order.",
         examples=[0, 1],
     )
-    sOrgOrdCancQty: int = Field(
+    sOrgOrdCancQty: float = Field(
         ...,
         title="원주문취소수량 (Original-order cancelled quantity)",
         description="Quantity cancelled from the original order.",
@@ -508,13 +508,13 @@ class AS0RealResponseBody(BaseModel):
         description="Order-user (placing employee) ID.",
         examples=["USER01"],
     )
-    sSpotOrdQty: int = Field(
+    sSpotOrdQty: float = Field(
         ...,
         title="실물주문수량 (Spot-order quantity)",
         description="Spot-order quantity in shares.",
         examples=[0, 1],
     )
-    sRuseOrdQty: int = Field(
+    sRuseOrdQty: float = Field(
         ...,
         title="재사용주문수량 (Reuse-order quantity)",
         description="Reuse-order quantity in shares.",
@@ -544,43 +544,43 @@ class AS0RealResponseBody(BaseModel):
         description="Commission consumed by the order.",
         examples=[0.0, 0.99],
     )
-    sSecBalQty: int = Field(
+    sSecBalQty: float = Field(
         ...,
         title="잔고수량 (Holding quantity)",
         description="Position holding quantity for the symbol after the event.",
         examples=[0, 10],
     )
-    sSpotOrdAbleQty: int = Field(
+    sSpotOrdAbleQty: float = Field(
         ...,
         title="실물주문가능수량 (Spot-orderable quantity)",
         description="Quantity available for spot ordering.",
         examples=[0, 10],
     )
-    sOrdAbleRuseQty: int = Field(
+    sOrdAbleRuseQty: float = Field(
         ...,
         title="주문가능재사용수량 (Reuse-orderable quantity)",
         description="Reusable quantity available for ordering.",
         examples=[0],
     )
-    sFlctQty: int = Field(
+    sFlctQty: float = Field(
         ...,
         title="변동수량 (Change quantity)",
         description="Net change in quantity caused by this event.",
         examples=[0, 1, -1],
     )
-    sSecBalQtyD2: int = Field(
+    sSecBalQtyD2: float = Field(
         ...,
         title="잔고수량(D2) (D+2 holding quantity)",
         description="D+2 settled holding quantity for the symbol.",
         examples=[0, 10],
     )
-    sSellAbleQty: int = Field(
+    sSellAbleQty: float = Field(
         ...,
         title="매도주문가능수량 (Sell-orderable quantity)",
         description="Quantity currently available to sell.",
         examples=[0, 10],
     )
-    sUnercSellOrdQty: int = Field(
+    sUnercSellOrdQty: float = Field(
         ...,
         title="미체결매도주문수량 (Unfilled sell-order quantity)",
         description="Sell-side quantity currently unfilled.",
